extract formatting of why-running
